requirements.txt
Flask
googlesearch-python
beautifulsoup4
requests
